Refactor: extract shared behavior
all action will have to check if the character is trying to do something in its camp or not
